Working store separation, now adjust file structure
