+++ src/main/js/Security.js
@@ -1,26 +1,31 @@
-// import { combineReducers } from 'redux';
 import 'whatwg-fetch';
+// For temporary page management
+import React from 'react';
+import { browserHistory } from 'react-router';
+import { push } from 'react-router-redux';
+
+import Utility from './util/Utility';
 
 // ----------------
 // DEFAULT STATE
-//
-
 export const defaultState = {
+  // Security
   loggedIn: false,
   loggedInUser: null,
   csrf: null
 };
 
 // -----------------
-// ACTIONS - These are serializable (hence replayable) descriptions of state transitions.
-// They do not themselves have any side-effects; they just describe something that is going to happen.
-// Use @typeName and isActionType for type detection that works even after serialization/deserialization.
-
+// ACTION CONSTANTS - Just used to differentiate the "actions"
 export const RECEIVE_CSRF = 'RECEIVE_CSRF';
 export const REQUEST_CSRF = 'REQUEST_CSRF';
 export const RECEIVE_LOGGED_IN_USER = 'RECEIVE_LOGGED_IN_USER';
 export const REQUEST_LOGGED_IN_USER = 'REQUEST_LOGGED_IN_USER';
 
+// -----------------
+// ACTION (CREATORS) - These are serializable (hence replayable) descriptions of state transitions.
+// They do not themselves have any side-effects; they just describe something that is going to happen.
+// Use @typeName and isActionType for type detection that works even after serialization/deserialization.
 export function receiveCSRF(csrf) {
   return {
     type: RECEIVE_CSRF,
@@ -52,9 +57,8 @@ export function requestLoggedInUser() {
 }
 
 export function fetchLoggedInUser() {
-  return function (dispatch) {
+  return function (dispatch, getState) {
     dispatch(requestLoggedInUser());
-
     fetch('/username', { credentials: 'same-origin' })
       .then(response => response.text())
       .then(user => dispatch(receiveLoggedInUser(user)));
@@ -86,10 +90,11 @@ export const actionCreators = {
 // ----------------
 // REDUCER - For a given state and action, returns the new state.
 // To support time travel, this must not mutate the old state.
-export const reducer = (state, action) => {
+export const security = (state, action) => {
   switch (action.type) {
     default:
       return state || defaultState;
+      // Security stuff
     case RECEIVE_CSRF:
       return {
         ...state,
@@ -114,7 +119,6 @@ export const reducer = (state, action) => {
   }
 };
 
-// Not needed since we only have 1 reducer
-// const randomStore = combineReducers({
-//   reducer
-// });
+export const reducers = {
+  security
+};
