+++ src/main/js/components/FourAlternativeQuestion.js
@@ -6,7 +6,7 @@ import { ButtonToolbar, Grid, Row, Col } from 'react-bootstrap';
 import AnswerButton from './AnswerButton';
 import Utility from '../util/Utility';
 
-import * as Store from '../Store';
+import * as Lessons from '../Lessons';
 
 export class FourAlternativeQuestion extends React.Component {
   constructor(props) {
@@ -228,7 +228,7 @@ FourAlternativeQuestion.propTypes = {
 // Wire up the React component to the Redux store and export it when importing this file
 export default connect(
     // Selects which state properties are merged into the component's props
-    state => (state.reducer),
+    state => (state.lessons),
     // Selects which action creators are merged into the component's props
-    Store.actionCreators
+    Lessons.actionCreators
 )(FourAlternativeQuestion);
