--- src/main/js/components/QuizPlayPage.js
@@ -1,6 +1,0 @@
-import React from 'react';
-import FourAlternativeQuestion from './FourAlternativeQuestion';
-
-const QuizPlayPage = () => <FourAlternativeQuestion pageName="QuizPlayPage" />;
-
-export default QuizPlayPage;
