--- src/main/js/store.old/index.js
@@ -1,14 +1,0 @@
-// // import * as WeatherForecasts from './WeatherForecasts';
-// // import * as Counter from './Counter';
-// import * as Store from './Store';
-
-// export const reducers = {
-//   userStatistics: UserStatistics.reducer,
-//   security: Security.reducer
-// };
-
-// export default {
-//   reducers() {
-//     return reducers;
-//   } // weatherForecasts: WeatherForecasts.reducer
-// };
